Separate sessions into submodules
