+++ apps/cli/commands/ai/sessions/helpers.ts
@@ -1,7 +1,8 @@
 import { select } from '@inquirer/prompts';
 import { __ } from '@wordpress/i18n';
 import chalk from 'chalk';
-import { listAiSessions, type AiSessionSummary } from 'cli/ai/sessions';
+import { listAiSessions } from 'cli/ai/sessions/store';
+import type { AiSessionSummary } from 'cli/ai/sessions/types';
 
 function formatSessionTimestamp( timestamp: string ): string {
 	const parsed = Date.parse( timestamp );
