+++ apps/cli/commands/ai/sessions/resume.ts
@@ -1,5 +1,5 @@
 import { __ } from '@wordpress/i18n';
-import { listAiSessions, loadAiSession } from 'cli/ai/sessions';
+import { listAiSessions, loadAiSession } from 'cli/ai/sessions/store';
 import { runCommand as runAiCommand } from 'cli/commands/ai';
 import { chooseSessionForAction } from 'cli/commands/ai/sessions/helpers';
 import { Logger, LoggerError } from 'cli/logger';
